WebDev/Blog/requirements.txt
alembic==1.15.2
bleach==6.2.0
blinker==1.9.0
boto3==1.38.2
botocore==1.38.2
certifi==2025.4.26
charset-normalizer==3.4.2
click==8.1.8
csscompressor==0.9.5
Deprecated==1.2.18
Flask==3.1.0
Flask-Caching==2.3.1
Flask-Compress==1.17
Flask-Limiter==3.12
Flask-Login==0.6.3
Flask-Mail==0.10.0
Flask-Migrate==4.1.0
Flask-SQLAlchemy==3.1.1
Flask-WTF==1.2.2
gunicorn==23.0.0
idna==3.10
itsdangerous==2.2.0
Jinja2==3.1.6
jmespath==1.0.1
jsmin==3.0.1
limits==5.2.0
Mako==1.3.10
Markdown==3.8
markdown-it-py==3.0.0
MarkupSafe==3.0.2
mdurl==0.1.2
ordered-set==4.1.0
packaging==25.0
pillow==11.2.1
psycopg2-binary==2.9.10
Pygments==2.19.1
python-dateutil==2.9.0.post0
python-dotenv==1.1.0
requests==2.32.3
rich==13.9.4
s3transfer==0.12.0
sentry-sdk[flask]==2.19.2
sib-api-v3-sdk==7.6.0
six==1.17.0
SQLAlchemy==2.0.40
typing_extensions==4.13.2
urllib3==2.4.0
webencodings==0.5.1
Werkzeug==3.1.3
wrapt==1.17.2
WTForms==3.2.1
